# src/api_service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API с JWT авторизацией",
    "description": "Пример API, использующего JWT для авторизации в Swagger UI",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/coworking/{coworking_id}/get": {
      "get": {
        "tags": [
          "Coworking"
        ],
        "summary": "Get Coworking",
        "operationId": "get_coworking_v1_coworking__coworking_id__get_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "coworking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Coworking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoworkingDTO"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Entity not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/coworking/list": {
      "get": {
        "tags": [
          "Coworking"
        ],
        "summary": "List Coworkings",
        "operationId": "list_coworkings_v1_coworking_list_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoworkingDTO"
                  },
                  "title": "Response List Coworkings V1 Coworking List Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/coworking/add": {
      "post": {
        "tags": [
          "Coworking"
        ],
        "summary": "Add Coworking",
        "operationId": "add_coworking_v1_coworking_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoworkingCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoworkingDTO"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/coworking/{coworking_id}/update": {
      "patch": {
        "tags": [
          "Coworking"
        ],
        "summary": "Update Coworking",
        "operationId": "update_coworking_v1_coworking__coworking_id__update_patch",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "coworking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Coworking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoworkingUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoworkingDTO"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Entity not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/coworking/{coworking_id}/delete": {
      "delete": {
        "tags": [
          "Coworking"
        ],
        "summary": "Delete Coworking",
        "operationId": "delete_coworking_v1_coworking__coworking_id__delete_delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "coworking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Coworking Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Entity not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/coworking/tariffs/add": {
      "post": {
        "tags": [
          "Coworking"
        ],
        "summary": "Add Tariffs",
        "operationId": "add_tariffs_v1_coworking_tariffs_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/CoworkingTariffCreateDTO"
                },
                "type": "array",
                "title": "Tariffs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CoworkingTariffDTO"
                  },
                  "type": "array",
                  "title": "Response Add Tariffs V1 Coworking Tariffs Add Post"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Access denied"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Entity not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/coworking/{coworking_id}/tariffs/list": {
      "get": {
        "tags": [
          "Coworking"
        ],
        "summary": "List Tariffs",
        "operationId": "list_tariffs_v1_coworking__coworking_id__tariffs_list_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "coworking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Coworking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoworkingTariffDTO"
                  },
                  "title": "Response List Tariffs V1 Coworking  Coworking Id  Tariffs List Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticate",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoworkingCreateDTO": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "examples": [
              "Коворкинг на Ленина"
            ]
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "title": "Address",
            "examples": [
              "г. Москва, ул. Ленина, 10"
            ]
          },
          "photo_url": {
            "type": "string",
            "minLength": 1,
            "title": "Photo Url",
            "examples": [
              "https://prod-team-37-ajc3mefd.REDACTED/api/v1/cdn/file/99216f07-ecb3-4b70-9fd6-1bbdae5d3104.jpeg"
            ]
          },
          "cover_url": {
            "type": "string",
            "minLength": 1,
            "title": "Cover Url",
            "examples": [
              "https://prod-team-37-ajc3mefd.REDACTED/api/v1/cdn/file/a078d506-a6e9-4f50-9a01-604c85ac4049.jpeg"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0,
                "examples": [
                  "Коворкинг на Ленина - это пространство для работы и общения"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "address",
          "photo_url",
          "cover_url"
        ],
        "title": "CoworkingCreateDTO"
      },
      "CoworkingDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "examples": [
              "487dec82-396e-4cf3-b555-86cc5db6b415"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "examples": [
              "Коворкинг на Ленина"
            ]
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "title": "Address",
            "examples": [
              "г. Москва, ул. Ленина, 10"
            ]
          },
          "photo_url": {
            "type": "string",
            "minLength": 1,
            "title": "Photo Url",
            "examples": [
              "https://prod-team-37-ajc3mefd.REDACTED/api/v1/cdn/file/99216f07-ecb3-4b70-9fd6-1bbdae5d3104.jpeg"
            ]
          },
          "cover_url": {
            "type": "string",
            "minLength": 1,
            "title": "Cover Url",
            "examples": [
              "https://prod-team-37-ajc3mefd.REDACTED/api/v1/cdn/file/a078d506-a6e9-4f50-9a01-604c85ac4049.jpeg"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0,
                "examples": [
                  "Коворкинг на Ленина - это пространство для работы и общения"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": [
              "2025-03-04T06:37:03.960093"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "photo_url",
          "cover_url",
          "created_at"
        ],
        "title": "CoworkingDTO"
      },
      "CoworkingTariffCreateDTO": {
        "properties": {
          "coworking_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Coworking Id",
            "examples": [
              "487dec82-396e-4cf3-b555-86cc5db6b415"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "examples": [
              "VIP"
            ]
          },
          "color": {
            "type": "string",
            "minLength": 1,
            "title": "Color",
            "examples": [
              "#4CA50F"
            ]
          },
          "price_per_hour": {
            "type": "integer",
            "minimum": 0,
            "title": "Price Per Hour",
            "examples": [
              750
            ]
          }
        },
        "type": "object",
        "required": [
          "coworking_id",
          "name",
          "color",
          "price_per_hour"
        ],
        "title": "CoworkingTariffCreateDTO"
      },
      "CoworkingTariffDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "examples": [
              "fe8279bc-2d9c-43ea-985c-0f3f53981335"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "examples": [
              "VIP"
            ]
          },
          "color": {
            "type": "string",
            "minLength": 1,
            "title": "Color",
            "examples": [
              "#4CA50F"
            ]
          },
          "price_per_hour": {
            "type": "integer",
            "minimum": 0,
            "title": "Price Per Hour",
            "examples": [
              750
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "examples": [
              "2025-03-04T06:37:03.960093"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color",
          "price_per_hour",
          "created_at"
        ],
        "title": "CoworkingTariffDTO"
      },
      "CoworkingUpdateDTO": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "examples": [
                  "Коворкинг на Ленина"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "examples": [
                  "г. Москва, ул. Ленина, 10"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "examples": [
                  "https://prod-team-37-ajc3mefd.REDACTED/api/v1/cdn/file/99216f07-ecb3-4b70-9fd6-1bbdae5d3104.jpeg"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "examples": [
                  "https://prod-team-37-ajc3mefd.REDACTED/api/v1/cdn/file/a078d506-a6e9-4f50-9a01-604c85ac4049.jpeg"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0,
                "examples": [
                  "Коворкинг на Ленина - это пространство для работы и общения"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "CoworkingUpdateDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен для авторизации. Формат: Bearer <токен>"
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}